backend_python/pyproject.toml
[project]
name = "backend-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.0",
    "supabase>=2.22.0",
    "uvicorn[standard]>=0.22.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "openai-agents>=0.3.3",
    "pymupdf>=1.26.5",
]
